Updated: Converted _payload.py to a submodule
